barChartのredux管理 #189
