+++ src/App.jsx
@@ -5,6 +5,7 @@ import { fetch_dots } from "./reducks/dots/action";
 import { fetch_todayDotLength } from "./reducks/star/action";
 import { fetch_icons } from "./reducks/userIcon/action";
 import { fetch_oneWeekDots } from "./reducks/oneWeekDots/action";
+import { fetch_lastWeekDots } from "./reducks/lastWeekDots/action";
 import Base from "./components/pages/Base";
 import Home from "./components/pages/Home";
 import SignIn from "./components/pages/SignIn";
@@ -138,7 +139,7 @@ export default function App() {
 					dispatch(fetch_todayDotLength(todayDot.length));
 				});
 
-			//-redux変更完了
+			//----今週の勉強時間(barChartで使用)----//
 			const array = [];
 			const specify_weekago = () => {
 				let agoDate = new Date();
@@ -151,34 +152,33 @@ export default function App() {
 				let four = three.setSeconds(0);
 				let five = new Date(four);
 				return five;
-      };
-      const init_arrayWeeks = () => {
-        const jsWeekAgo = [];
-        let today = new Date();
-        today.setDate(today.getDate() + 1);
-        const infoWeek = [];
-        const infoDay = [];
-        const subtract = 1;
-        const max = 6;
-        for (let i = 0; i <= max; i++) {
-          today.setDate(today.getDate() - subtract);
-          infoWeek[i] = today.getMonth() + 1 + "/" + today.getDate();
-          infoDay[i] = today.getDay();
-          jsWeekAgo.push({
-            label: infoWeek[i],
-            jsGetDay: infoDay[i],
-            initNum: 0,
-          });
-        }
-        const reversedLabelWeek = infoWeek.reverse();
-        const reversedDataWeek = jsWeekAgo.reverse();
-        const weeksObj = {
-          weekLabel: reversedLabelWeek,
-          weekData: reversedDataWeek,
-        };
-        return weeksObj;
-      };
-
+			};
+			const init_arrayWeeks = () => {
+				const jsWeekAgo = [];
+				let today = new Date();
+				today.setDate(today.getDate() + 1);
+				const infoWeek = [];
+				const infoDay = [];
+				const subtract = 1;
+				const max = 6;
+				for (let i = 0; i <= max; i++) {
+					today.setDate(today.getDate() - subtract);
+					infoWeek[i] = today.getMonth() + 1 + "/" + today.getDate();
+					infoDay[i] = today.getDay();
+					jsWeekAgo.push({
+						label: infoWeek[i],
+						jsGetDay: infoDay[i],
+						initNum: 0,
+					});
+				}
+				const reversedLabelWeek = infoWeek.reverse();
+				const reversedDataWeek = jsWeekAgo.reverse();
+				const weeksObj = {
+					weekLabel: reversedLabelWeek,
+					weekData: reversedDataWeek,
+				};
+				return weeksObj;
+			};
 			firebase
 				.firestore()
 				.collection("dots")
@@ -211,9 +211,93 @@ export default function App() {
 					});
 					dispatch(fetch_oneWeekDots(finalWeek));
 				});
+			// -----------------------------
+
+			// ------Last weekの勉強時間-------
+			let yesterday2 = new Date(
+				specify_weekago().getFullYear(),
+				specify_weekago().getMonth(),
+				specify_weekago().getDate() - 1
+			);
+			const yesterday = new Date(yesterday2.setHours(23, 59, 59, 999));
+			//前週を指定
+			const specify_lastweek = () => {
+				let agoWeek2 = yesterday2.setDate(yesterday2.getDate() - 6);
+				let hope2 = new Date(agoWeek2);
+				let zero2 = hope2.setHours(0);
+				let one2 = new Date(zero2);
+				let two2 = one2.setMinutes(0);
+				let three2 = new Date(two2);
+				let four2 = three2.setSeconds(0);
+				let five2 = new Date(four2);
+				return five2;
+			};
+			//前週の日付とラベルを取得
+			const init_arrayWeeks2 = () => {
+				let today = new Date(specify_weekago());
+				let yesterday = new Date(
+					today.getFullYear(),
+					today.getMonth(),
+					today.getDate() - 1
+				);
+				const jsWeekAgo2 = [];
+				yesterday.setDate(yesterday.getDate() + 1);
+				const infoWeek = [];
+				const infoDay = [];
+				const subtract = 1;
+				const max = 6;
+				for (let i = 0; i <= max; i++) {
+					yesterday.setDate(yesterday.getDate() - subtract);
+					infoWeek[i] = yesterday.getMonth() + 1 + "/" + yesterday.getDate();
+					infoDay[i] = yesterday.getDay();
+					jsWeekAgo2.push({
+						label: infoWeek[i],
+						jsGetDay: infoDay[i],
+						initNum: 0,
+					});
+				}
+				const reversedLabelWeek = infoWeek.reverse();
+				const reversedDataWeek = jsWeekAgo2.reverse();
+				const weeksObj = {
+					weekLabel: reversedLabelWeek,
+					weekData: reversedDataWeek,
+				};
+				return weeksObj;
+			};
+			//firestoreのDBでも時間が指定できるように変換する
+			const startDate = firebase.firestore.Timestamp.fromDate(
+				specify_lastweek()
+			);
+			const endDate = firebase.firestore.Timestamp.fromDate(yesterday);
+			firebase
+				.firestore()
+				.collection("dots")
+				.orderBy("createdAt")
+				.where("userId", "==", currentUser.uid)
+				.startAt(startDate)
+				.endBefore(endDate)
+				.get()
+				.then((snapshot) => {
+					const hope2 = init_arrayWeeks2().weekData;
+					snapshot.docs.map((doc) => {
+						const item2 = doc.data();
+						const receivedDay = item2.getday;
+						const hours = item2.working;
+						for (let i = 0; i < hope2.length; i++) {
+							if (receivedDay === hope2[i].jsGetDay) {
+								hope2[i].initNum = hours;
+							}
+						}
+					});
+					const finalWeek2 = hope2.map((el) => {
+						return el.initNum;
+					});
+					dispatch(fetch_lastWeekDots(finalWeek2))
+					// set_filledWeek2(finalWeek2);
+				});
 		}
 	}, [currentUser]);
-
+	// ----------------
 
 	return (
 		<>
